refactor(pom-cli): 入力読み込みと pptx-glimpse フォント設定を共通モジュールに抽出
build.ts / preview.ts に重複していた md/xml 読み込み・masterPptx 解決を
input.ts に、バンドルフォント解決と追加フォントマッピングを glimpse.ts に
切り出す。render サブコマンド追加 (#816) の下準備。
